Changed gitignore and did cleanup
added vs files to gitignore,
combined all of the enviornment items into one object called Level, made it into a prefab, and made copies to the left and right.

combined the camera and character marker into the Player object to make it easier to move the starting area without having to move the camera starting area as well.
Changed some minor camera settings as well.